# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  ActiveSessionListResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/ActiveSessionResponse'
        type: array
      total:
        type: integer
    required:
    - results
    - total
    type: object
  ActiveSessionListResponseDoc:
    properties:
      data:
        $ref: '#/definitions/ActiveSessionListResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  ActiveSessionResponse:
    properties:
      browserName:
        type: string
      cityName:
        type: string
      clientIP:
        type: string
      countryCode:
        type: string
      createdAt:
        type: string
      current:
        type: boolean
      deviceLabel:
        type: string
      deviceName:
        type: string
      deviceType:
        type: string
      expiresAt:
        type: string
      geoAccuracy:
        type: string
      geoSource:
        type: string
      ipLatitude:
        type: number
        x-nullable: true
        x-omitempty: false
      ipLongitude:
        type: number
        x-nullable: true
        x-omitempty: false
      lastSeenAt:
        type: string
        x-nullable: true
        x-omitempty: false
      locationLabel:
        type: string
      osName:
        type: string
      preciseAccuracyMeters:
        type: number
        x-nullable: true
        x-omitempty: false
      preciseLatitude:
        type: number
        x-nullable: true
        x-omitempty: false
      preciseLocatedAt:
        type: string
        x-nullable: true
        x-omitempty: false
      preciseLongitude:
        type: number
        x-nullable: true
        x-omitempty: false
      regionName:
        type: string
      sessionID:
        type: string
      timezoneName:
        type: string
      updatedAt:
        type: string
    required:
    - browserName
    - cityName
    - clientIP
    - countryCode
    - createdAt
    - current
    - deviceLabel
    - deviceName
    - deviceType
    - expiresAt
    - geoAccuracy
    - geoSource
    - ipLatitude
    - ipLongitude
    - lastSeenAt
    - locationLabel
    - osName
    - preciseAccuracyMeters
    - preciseLatitude
    - preciseLocatedAt
    - preciseLongitude
    - regionName
    - sessionID
    - timezoneName
    - updatedAt
    type: object
  ActiveSessionResponseDoc:
    properties:
      data:
        $ref: '#/definitions/ActiveSessionResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  AuthErrorDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        example: invalid_request
        type: string
      errorMsg:
        example: invalid request
        type: string
      requestId:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  AuthUserResponse:
    properties:
      appearancePreferences:
        type: string
      avatarURL:
        type: string
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: integer
      lastActiveAt:
        type: string
        x-nullable: true
        x-omitempty: false
      lastLoginAt:
        type: string
        x-nullable: true
        x-omitempty: false
      locale:
        type: string
      profilePreferences:
        type: string
      publicID:
        type: string
      role:
        type: string
      status:
        type: string
      timezone:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    required:
    - appearancePreferences
    - avatarURL
    - createdAt
    - displayName
    - email
    - id
    - lastActiveAt
    - lastLoginAt
    - locale
    - profilePreferences
    - publicID
    - role
    - status
    - timezone
    - updatedAt
    - username
    type: object
...
  ChangePasswordRequest:
    properties:
      currentPassword:
        maxLength: 128
        type: string
      newPassword:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
  ChangePasswordResponse:
    properties:
      changed:
        type: boolean
    required:
    - changed
    type: object
  ChangePasswordResponseDoc:
    properties:
      data:
        $ref: '#/definitions/ChangePasswordResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  EmailRegistrationCompleteRequest:
    properties:
      code:
        type: string
      email:
        maxLength: 128
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
      turnstileToken:
        maxLength: 2048
        type: string
    required:
    - email
    - password
    type: object
  EmailRegistrationStartRequest:
    properties:
      email:
        maxLength: 128
        type: string
      turnstileToken:
        maxLength: 2048
        type: string
    required:
    - email
    type: object
  EmailRegistrationStartResponse:
    properties:
      expiresAt:
        type: string
      sent:
        type: boolean
    required:
    - expiresAt
    - sent
    type: object
  EmailRegistrationStartResponseDoc:
    properties:
      data:
        $ref: '#/definitions/EmailRegistrationStartResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  LoginOptionsResponse:
    properties:
      emailRegistrationEnabled:
        type: boolean
      emailVerificationEnabled:
        type: boolean
      turnstileEnabled:
        type: boolean
      turnstileSiteKey:
        type: string
    required:
    - emailRegistrationEnabled
    - emailVerificationEnabled
    - turnstileEnabled
    - turnstileSiteKey
    type: object
  LoginOptionsResponseDoc:
    properties:
      data:
        $ref: '#/definitions/LoginOptionsResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  LoginRequest:
    properties:
      email:
        maxLength: 128
        type: string
      password:
        maxLength: 128
        minLength: 6
        type: string
      turnstileToken:
        maxLength: 2048
        type: string
    required:
    - email
    - password
    type: object
  LoginResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
        x-nullable: true
      refreshExpiresAt:
        type: string
        x-nullable: true
      sessionID:
        type: string
      twoFactorChallengeToken:
        type: string
      twoFactorRequired:
        type: boolean
    required:
    - twoFactorRequired
    type: object
  LoginResponseDoc:
    properties:
      data:
        $ref: '#/definitions/LoginResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  LogoutResponse:
    properties:
      revoked:
        type: boolean
    required:
    - revoked
    type: object
  LogoutResponseDoc:
    properties:
      data:
        $ref: '#/definitions/LogoutResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  MeResponse:
    properties:
      user:
        $ref: '#/definitions/AuthUserResponse'
    required:
    - user
    type: object
  MeResponseDoc:
    properties:
      data:
        $ref: '#/definitions/MeResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  PatchMeRequest:
    properties:
      appearancePreferences:
        maxLength: 65536
        type: string
      avatarURL:
        maxLength: 2048
        type: string
      displayName:
        maxLength: 128
        type: string
      locale:
        maxLength: 32
        type: string
      profilePreferences:
        maxLength: 65536
        type: string
      timezone:
        maxLength: 64
        type: string
    type: object
...
  TwoFactorVerifyRequest:
    properties:
      challengeToken:
        maxLength: 4096
        minLength: 20
        type: string
      code:
        maxLength: 32
        minLength: 6
        type: string
    required:
    - challengeToken
    - code
    type: object
...
  UpdateCurrentSessionLocationRequest:
    properties:
      accuracyMeters:
        maximum: 100000
        minimum: 0
        type: number
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      timezone:
        maxLength: 64
        type: string
    required:
    - latitude
    - longitude
    type: object
...
info:
  contact: {}
  description: DEEIX Chat 后端 API 文档
  title: DEEIX Chat API
  version: "0.4.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates an email/password with Sub2 and sets the DEEIX refresh
        cookie on success.
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      summary: Log in with Sub2
      tags:
      - auth
  /auth/login-options:
    get:
      description: Returns current Sub2 email registration and Turnstile settings.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginOptionsResponseDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      summary: Get Sub2 login options
      tags:
      - auth
  /auth/login/2fa:
    post:
      consumes:
      - application/json
      description: Verifies the encrypted Sub2 login challenge and sets the DEEIX
        refresh cookie on success.
      parameters:
      - description: Sub2 login 2FA request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/TwoFactorVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      summary: Complete Sub2 login 2FA
      tags:
      - auth
  /auth/logout:
    post:
      description: Revokes the current DEEIX session and clears its refresh cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LogoutResponseDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      security:
      - BearerAuth: []
      summary: Log out the current browser session
      tags:
      - auth
  /auth/logout-all:
    post:
      description: Revokes all DEEIX sessions for the current Sub2 principal and clears
        the refresh cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LogoutResponseDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      security:
      - BearerAuth: []
      summary: Log out all browser sessions
      tags:
      - auth
  /auth/refresh:
    post:
      description: Rotates the DEEIX session using only the HttpOnly refresh cookie
        and returns no refresh token in JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginResponseDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      summary: Refresh the browser session
      tags:
      - auth
  /auth/register/email/complete:
    post:
      consumes:
      - application/json
      description: Registers with Sub2 and creates a DEEIX browser session with a
        refresh cookie.
      parameters:
      - description: Registration completion
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/EmailRegistrationCompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      summary: Complete Sub2 email registration
      tags:
      - auth
  /auth/register/email/start:
    post:
      consumes:
      - application/json
      description: Requests a Sub2 registration code for an email address.
      parameters:
      - description: Registration request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/EmailRegistrationStartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EmailRegistrationStartResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      summary: Start Sub2 email registration
      tags:
      - auth
  /auth/sessions:
    get:
      description: Lists active DEEIX sessions for the current Sub2 principal.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ActiveSessionListResponseDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      security:
      - BearerAuth: []
      summary: List active browser sessions
      tags:
      - auth
  /auth/sessions/{session_id}/logout:
    post:
      description: Revokes the specified DEEIX browser session and clears its stored
        Sub2 credentials.
      parameters:
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LogoutResponseDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      security:
      - BearerAuth: []
      summary: Revoke a browser session
      tags:
      - auth
  /auth/sessions/current/location:
    put:
      consumes:
      - application/json
      description: Updates DEEIX session metadata for the authenticated browser session.
      parameters:
      - description: Current session location
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateCurrentSessionLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ActiveSessionResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      security:
      - BearerAuth: []
      summary: Update current session location
      tags:
      - auth
...
  /me:
    get:
      description: Returns the revalidated Sub2 principal and DEEIX-owned preferences.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MeResponseDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      security:
      - BearerAuth: []
      summary: Get the current Sub2 principal projection
      tags:
      - auth
    patch:
      consumes:
      - application/json
      description: Updates DEEIX-owned display, avatar, locale, timezone, and preference
        fields.
      parameters:
      - description: Profile update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/PatchMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MeResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      security:
      - BearerAuth: []
      summary: Update local profile preferences
      tags:
      - auth
  /me/password:
    put:
      consumes:
      - application/json
      description: Proxies the password change to Sub2, then revokes DEEIX browser
        sessions.
      parameters:
      - description: Password change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ChangePasswordResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/AuthErrorDoc'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/AuthErrorDoc'
      security:
      - BearerAuth: []
      summary: Change the Sub2 password
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"